refactor(pi): split the engine-neutral turn mechanism out of the harness L0
The AgentSession L0 was importing the harness L0's internals - lease, terminals,
event queue, prompt prep - and two comments existed only to explain why those
were exported. The dependency was backwards: none of those parts care which pi
class runs the turn.

They move to turn-kit.ts, which both L0s now sit on. invoke.ts drops from 721 to
510 lines and its header says what it is: the harness L0 plus the event
vocabulary and observation plane only it has.

The cancellation protocol's two halves also stop being copied per L0:
cancellableStream joins abortFirstIterator in collect.ts, so a turn generator
receives the latch and publishes its door through one shared wrapper.

No behaviour change.
